新增加接口信息表的各层实现

+++ api-backend/src/main/java/com/fusheng/api_backend/model/vo/sysUser/SysUserInfoVO.java
@@ -50,6 +50,6 @@ public class SysUserInfoVO {
     /**
      * 状态
      */
-    private Byte userStatus;
+    private Integer userStatus;
 
 }
